feature/GH-14393: move cart validation service and guard from component to core
