+++ feature-libs/cart/main/components/cart/cart-details/cart-details.component.ts
@@ -1,4 +1,5 @@
 import { ChangeDetectionStrategy, Component, OnInit } from '@angular/core';
+import { CartConfigService } from '@spartacus/cart/main/core';
 import {
   ActiveCartFacade,
   Cart,
@@ -9,7 +10,6 @@ import {
 import { AuthService, RoutingService } from '@spartacus/core';
 import { combineLatest, Observable, of } from 'rxjs';
 import { filter, map, tap } from 'rxjs/operators';
-import { CartConfigService } from '../../services/cart-config.service';
 
 @Component({
   selector: 'cx-cart-details',
