+++ feature-libs/cart/main/components/cart/index.ts
@@ -14,7 +14,5 @@ export * from './order-entries-context/index';
 export * from './save-for-later/save-for-later.component';
 export * from './save-for-later/save-for-later.module';
 export * from './validation/cart-validation-components.module';
-export * from './validation/cart-validation-state.service';
 export * from './validation/cart-warnings/cart-validation-warnings.component';
 export * from './validation/cart-warnings/cart-validation-warnings.module';
-export * from './validation/guards/index';
