+++ feature-libs/cart/main/components/index.ts
@@ -1,3 +1,2 @@
 export * from './cart-components.module';
 export * from './cart/index';
-export * from './services/index';
