--- feature-libs/cart/main/components/services/index.ts
@@ -1,1 +1,0 @@
-export * from './cart-config.service';
