+++ feature-libs/cart/main/core/index.ts
@@ -3,6 +3,7 @@ export * from './cart-persistence.module';
 export * from './connectors/index';
 export * from './event/index';
 export * from './facade/index';
+export * from './guards/index';
 export * from './services/index';
 export * from './store/actions/index';
 export * from './store/multi-cart-state';
