+++ feature-libs/cart/main/core/services/index.ts
@@ -1,1 +1,3 @@
+export * from './cart-config.service';
+export * from './cart-validation-state.service';
 export * from './multi-cart-state-persistence.service';
