+++ feature-libs/checkout/components/components/cost-center/cost-center.module.ts
@@ -1,6 +1,6 @@
 import { CommonModule } from '@angular/common';
 import { NgModule } from '@angular/core';
-import { CartValidationGuard } from '@spartacus/cart/main/components';
+import { CartValidationGuard } from '@spartacus/cart/main/core';
 import { CmsConfig, ConfigModule, I18nModule } from '@spartacus/core';
 import { CartNotEmptyGuard } from '../../guards/cart-not-empty.guard';
 import { CheckoutAuthGuard } from '../../guards/checkout-auth.guard';
