+++ feature-libs/checkout/components/components/payment-type/payment-type.module.ts
@@ -1,6 +1,6 @@
 import { CommonModule } from '@angular/common';
 import { NgModule } from '@angular/core';
-import { CartValidationGuard } from '@spartacus/cart/main/components';
+import { CartValidationGuard } from '@spartacus/cart/main/core';
 import { CmsConfig, ConfigModule, I18nModule } from '@spartacus/core';
 import { SpinnerModule } from '@spartacus/storefront';
 import { CartNotEmptyGuard } from '../../guards/cart-not-empty.guard';
