+++ feature-libs/checkout/components/components/shipping-address/shipping-address.module.ts
@@ -1,7 +1,7 @@
 import { CommonModule } from '@angular/common';
 import { NgModule } from '@angular/core';
 import { RouterModule } from '@angular/router';
-import { CartValidationGuard } from '@spartacus/cart/main/components';
+import { CartValidationGuard } from '@spartacus/cart/main/core';
 import { CmsConfig, I18nModule, provideDefaultConfig } from '@spartacus/core';
 import {
   AddressFormModule,
